feat: Add get method to Dictionary class
This commit adds a new `get` method to the `Dictionary` class in the `UtilsCPP` module. The `get` method allows retrieving the value associated with a given key from the dictionary. If the key is not present in the dictionary, the method inserts a default value and returns it.

The `get` method improves the usability and convenience of the `Dictionary` class by providing a simple way to access values while ensuring that the dictionary always contains the requested key.

This change addresses the need for a more efficient and concise way to retrieve values from the dictionary, enhancing the overall functionality of the `Dictionary` class.